moved redirect step to correct folder
